// resources/js/support/pagination.ts
export type PaginationLink = {
    url: string | null;
    label: string;
    active: boolean;
};

export type Paginated<T> = {
    data: T[];
    current_page?: number;
    last_page: number;
    from?: number | null;
    to?: number | null;
    total?: number;
    per_page?: number;
    links: PaginationLink[];
};